% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_from_DOY.R
\name{date_from_DOY}
\alias{date_from_DOY}
\title{date from DOY}
\usage{
date_from_DOY(decimal_day, origin, tz = "UTC")
}
\arguments{
\item{decimal_day}{date stored as DOY.decimalday}

\item{origin}{origin of decimal_day as YYYY/MM/DD}

\item{tz}{timezone}
}
\value{
decimal_day as POSIXct
}
\description{
converts DOY.decimalday into a POSIXct object. - note as this is DOY, Jan 1st 1200 is 1.5
}
