% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_date_from_day.R
\name{decimal_day_from_date}
\alias{decimal_day_from_date}
\title{Decimal day from date}
\usage{
decimal_day_from_date(date)
}
\arguments{
\item{date}{a POSIXct date}
}
\description{
convert a POSIXct in to a number representing day.decimalday, where Jan 1st 12:00 == 1.5
}
\author{
W. S. Drysdale
}
