% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOY_from_date.R
\name{DOY_from_date}
\alias{DOY_from_date}
\title{DOY from date}
\usage{
DOY_from_date(date)
}
\arguments{
\item{date}{a POSIXct date}
}
\description{
convert a POSIXct in to a number representing DOY, where Jan 1st 12:00 == 1.5
}
\author{
W. S. Drysdale
}
