% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_time_series.R
\name{pad_time_series}
\alias{pad_time_series}
\title{Pad Time Series}
\usage{
pad_time_series(.data, period)
}
\arguments{
\item{.data}{data.frame or tibble with a column called "date" of class POSIXct}

\item{period}{numeric value to be passed to \code{seq()} to create time series to be joined onto.}
}
\description{
Pad a time series by joining it to a regularly spaced sequence of dates.
}
\author{
W. S. Drysdale
}
