% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rle.R
\name{tidy_rle}
\alias{tidy_rle}
\title{tidy rle}
\usage{
tidy_rle(rle_obj)
}
\arguments{
\item{rle_obj}{output of \code{rle()}}
}
\description{
Take the output of \code{rle()}, convert to a tibble and calculate start and end indices.
}
\author{
W. S. Drysdale
}
