% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nasa_ames.R
\name{write_nasa_ames}
\alias{write_nasa_ames}
\title{Create NASA Ames 1D}
\usage{
write_nasa_ames(
  d,
  ONAME,
  ORG,
  SNAME,
  MNAME,
  IVOL = 1,
  NVOL = 1,
  DATE,
  DX,
  XNAME,
  X_col = 1,
  NV_col,
  VSCALn = NULL,
  VMISSn = NULL,
  VNAMEn,
  SCOMn = NULL,
  NCOMn = NULL,
  file,
  col.names = FALSE
)
}
\arguments{
\item{d}{data frame of data to have header attached, column names should be simple, no need to match long names in header}

\item{ONAME}{string containing lastname, firstname of file originator max length 132 characters}

\item{ORG}{string containing name of organisation max length 132 characters}

\item{SNAME}{string containing name of instrumentation max length 132 characters}

\item{MNAME}{string containing name of mission/campaign/project max length 132 characters}

\item{IVOL}{integer This files volume index}

\item{NVOL}{integer Total number of files in volume (i.e. files that share ONAME, ORG, SNAME and MNAME)}

\item{DATE}{string of date when data was taken in format to be parsed by \code{lubridate::ymd()}}

\item{DX}{interval of independent variable, zero if not consistent}

\item{XNAME}{string containing name of independent variable max length 132 characters}

\item{X_col}{column number containing independent variable}

\item{NV_col}{integer or vector of integers corresponding to column number containing primary variables}

\item{VSCALn}{integer or vector of integers corresponding to scaling factors of primary variables, default all 1}

\item{VMISSn}{integer or vector of integers corresponding to missing values of primary variables, default all 99999}

\item{VNAMEn}{string or vector of string containing long names of primary variables, each should not exceed 132 characters}

\item{SCOMn}{string or vector of string containing each special comment line}

\item{NCOMn}{string or vector of string containing each normal comment line}

\item{file}{save file here}

\item{col.names}{include the column headers on the data table T/F. If true they will be added as the final normal comments line}
}
\description{
{
Creates a 1 Dimentional NASA Ames file in the format \cr
\cr
\code{HL FFI} Header Length -  File Format Index (supports 1001 only) \cr
\code{ONAME} Name of File Originator \cr
\code{ORG} Name of Organisation \cr
\code{SNAME} Name of Instrumentation \cr
\code{MNAME} Campagin Name \cr
\code{IVOL NVOL} Data Volume number - total number of volumes \cr
\code{DATE RDATE} Date of data collection - data of file creation \cr
\code{DX} Interval in independant variable, 0 if interval is not consistant
\code{XNAME} Name of independant variable \cr
\code{NV} number of primary variables  \cr
\code{VSCAL1...VSCALn} scaling factors for variables \cr
\code{VMISS1...VMISSn} missing flag for variables \cr
\code{VNAME1} first long name of variables \cr
\code{...}  \cr
\code{VNAMEn} last long name of variables \cr
\code{NSCOML} number of special comment lines \cr
\code{SCOM1} first special comment line \cr
\code{...} \cr
\code{SCOMn} last special comment line \cr
\code{NNCOML} number of normal comment lines \cr
\code{NCOM1} first normal comment line \cr
\code{...} \cr
\code{NCOMn} last normal comment line \cr
}
}
\author{
W. S. Drysdale
}
