% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsdmiscr2-package.R
\docType{package}
\name{wsdmiscr2-package}
\alias{wsdmiscr2}
\alias{wsdmiscr2-package}
\title{wsdmiscr2: wsdmiscr2}
\description{
Miscelleneous functions useful in atmospheric science research.
}
\author{
\strong{Maintainer}: Will Drysdale \email{will.drysdale@york.ac.uk}

}
\keyword{internal}
