% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_from_decimal_day.R
\name{date_from_decimal_day}
\alias{date_from_decimal_day}
\title{date_from_decimal_day}
\usage{
date_from_decimal_day(decimal_day, date_origin, tz = "UTC")
}
\arguments{
\item{decimal_day}{date stored as DOY.decimalday}

\item{date_origin}{origin of decimal_day as YYYY/MM/DD}

\item{tz}{timezone}
}
\value{
decimal_day as POSIXct
}
\description{
converts DOY.decimalday into a POSIXct object.
}
